added: program for converting kilometers to miles

+++ basic-programs/AddTwoNum.java
@@ -3,10 +3,10 @@ public class AddTwoNum {
     public static void main(String[] args){
         System.out.println("Enter first number: ");
         Scanner input = new Scanner(System.in);
-        int num1 = input.nextInt();
+        double num1 = input.nextDouble();
         System.out.println("Enter second number: ");
-        int num2 = input.nextInt();
-        int sum = num1 + num2;
+        double num2 = input.nextDouble();
+        double sum = num1 + num2;
         System.out.println("The sum of first number and second number is: " + sum);
         input.close();
     }
